Add behavior to league class
The behavior standings is to sort through the team object and order
the teams by points in decending order.

The playoff_teams method is to sort the top half of the teams.

The position_for method is to find where each team are rank base
on the team being search.